interview_speech_module/requirements.txt
fastapi
uvicorn
vosk
librosa
numpy
scipy
praat-parselmouth
soundfile
